Remove plot images
